+++ Source/Foundation/NSString+SDExtensions.m
@@ -245,13 +245,13 @@
     for (NSInteger i = 0; i < patterns.count; i++)
     {
         NSString *currentMatch = [match stringByAppendingString:@"(\\d+)"];
-        match = [match stringByAppendingString:[NSString stringWithFormat:@"(\\d{%ld})", (long)((NSNumber *)[patterns objectAtIndex:i]).integerValue]];
+        match = [match stringByAppendingString:[NSString stringWithFormat:@"(\\d{%zd})", ((NSNumber *)[patterns objectAtIndex:i]).integerValue]];
 
         NSString *template;
         if (i == 0)
-            template = [NSString stringWithFormat:@"$%li", (long)i+1];
+            template = [NSString stringWithFormat:@"$%zd", i+1];
         else
-            template = [NSString stringWithFormat:@"%@$%li", [separators objectAtIndex:(NSUInteger)i-1], (long)i+1];
+            template = [NSString stringWithFormat:@"%@$%zd", [separators objectAtIndex:(NSUInteger)i-1], i+1];
 
         replace = [replace stringByAppendingString:template];
         [expressions addObject:@{@"match": currentMatch, @"replace": replace}];
